Restructured project - Moved api config to config folder - Moved axios config to config folder
