+++ jobs/Frontend/search-movies/src/components/Search/index.tsx
@@ -2,8 +2,8 @@ import _ from "lodash";
 import { useState } from "react";
 import { FaSearch } from "react-icons/fa";
 import styled from "styled-components";
-import { API_KEY, endpoints } from "../../utils/api";
-import { axiosBrowse } from "../../utils/axios";
+import { API_KEY, endpoints } from "../../config/api";
+import { axiosBrowse } from "../../config/axios";
 import { colors } from "../../utils/theme";
 import { shadowInner } from "../General";
 import SearchResults from "./SearchResults";
